Adjust the UI for mobile device screens
Added media queries to the stylesheet so users can play the game from
their mobile devices. When placing ships on the board from a mobile
device the last ship location was still highlighted due to the hover
selector being disabled. This allowed the opponent to see the player's
last ship location. In order to fix this bug on mobile devices
a call to mouseOutEvent() was made to remove the highlight class after
all ships were placed.